# specs/1-backend-fastapi-todo/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App API
  description: Backend API for Todo App with FastAPI and SQLModel
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User identifier
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '404':
          description: User not found
    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User identifier
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Task identifier
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User identifier
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Task identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User identifier
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Task identifier
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          description: Task not found
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User identifier
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Task identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Task identifier
        title:
          type: string
          description: Task title
          maxLength: 255
        description:
          type: string
          description: Task description
          maxLength: 1000
        completed:
          type: boolean
          description: Whether the task is completed
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
        tags:
          type: array
          items:
            type: string
          description: Task tags
        due_date:
          type: string
          format: date-time
          description: Task due date
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Task update timestamp
        user_id:
          type: string
          description: User identifier
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
          maxLength: 255
        description:
          type: string
          description: Task description
          maxLength: 1000
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          description: Task priority level
        tags:
          type: array
          items:
            type: string
          description: Task tags
        due_date:
          type: string
          format: date-time
          description: Task due date
        completed:
          type: boolean
          default: false
          description: Whether the task is completed
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Task title
          maxLength: 255
        description:
          type: string
          description: Task description
          maxLength: 1000
        completed:
          type: boolean
          description: Whether the task is completed
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
        tags:
          type: array
          items:
            type: string
          description: Task tags
        due_date:
          type: string
          format: date-time
          description: Task due date